Updated Field "centerering" from logical to float
This commit changes "centering" in Fields from an ill-defined "bool"
value (does true mean lower face or upper face?) to a numerical offset
from the center of a cell, with values in the range [-1.0,+1.0] for
each axis.  Active array sizes are unchanged as they should be, but
application code may require an extra ghost zone.  Below summarizes
the main changes to the Field API and parameters.

Field API changes

   OLD

      void set_centering(int id, int cx, int cy=0, int cz=0)
      void centering(int id, int * cx, int * cy = 0, int * cz = 0) const

   NEW

      void set_centering(int id, double cx,   double cy,   double cz) throw()
      void get_centering(int id, double & cx, double & cy, double & cz) const

Additionally, the "Field : <field> : centering" parameter type
(http://client65-90.sdsc.edu/cello-doc/parameters-list.html#field) is
changed from "list (logical)" to "list (float)", so parameter files
that use the centering parameter will need to be updated.  Range checking
is done when the parameter is read in.

Unit tests in test_Field, test_FieldData and test_FieldDescr have been
updated accordingly.

+++ src/Cello/parameters_Config.hpp
@@ -376,7 +376,7 @@ public: // attributes
   std::vector<std::string>   field_list;
   std::map<std::string,int>  field_index;
   int                        field_alignment;
-  std::vector<int>           field_centering [3];
+  std::vector<double>        field_centering [3];
   int                        field_ghost_depth[3];
   int                        field_padding;
   int                        field_history;
